functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin==6.2.0
python-dotenv==1.0.1
requests==2.32.3
google-cloud-firestore==2.19.0
google-cloud-logging==3.11.3
google-cloud-monitoring==2.26.0
earthengine-api==1.4.3
pillow==11.0.0
Flask-Cors==5.0.0
